refactoring widgets: creation of CoronedCard, rename others
- decompose cards to have a custom widget with just card and possibility to put children in a Column
