+++ lib/ui/widgets/widgets.dart
@@ -1,5 +1,6 @@
 export 'navbar_item.dart';
 export 'navigation_drawer/navigation_drawer.dart';
-export 'coroned_global_card.dart';
-export 'coroned_country_card.dart';
+export 'cards/global_card.dart';
+export 'cards/country_card.dart';
+export 'cards/coroned_card.dart';
 export 'failure_icon.dart';
